214: Rename reddit_fetcher to fetcher
