+++ backend/src/api/report/generate.rs
@@ -1,15 +1,15 @@
 use crate::api::report::report_ack::ReportAck;
 use crate::app_error::AppError;
 use crate::auth::google::{GoogleId, JwtUserInfo};
+use crate::fetcher::feed_request::{FetcherFeedRequest, RedditFeedKind};
+use crate::fetcher::fetcher::RMoodsFetcher;
+use crate::fetcher::model::post_comments::PostComments;
+use crate::fetcher::model::posts::Posts;
+use crate::fetcher::model::reddit_data::RedditFeedData;
+use crate::fetcher::model::user_posts::UserPosts;
 use crate::nlp::analysis::NlpAnalysisKind;
 use crate::nlp::nlp_client::NlpClient;
 use crate::nlp::report::{new_report_id, RMoodsReport, ReportMetadata};
-use crate::reddit_fetcher::feed_request::{FetcherFeedRequest, RedditFeedKind};
-use crate::reddit_fetcher::fetcher::RMoodsFetcher;
-use crate::reddit_fetcher::model::post_comments::PostComments;
-use crate::reddit_fetcher::model::posts::Posts;
-use crate::reddit_fetcher::model::reddit_data::RedditFeedData;
-use crate::reddit_fetcher::model::user_posts::UserPosts;
 use crate::websocket::SystemMessage;
 use crate::websocket::SystemMessage::ReportError;
 use crate::AppState;
