+++ backend/src/api/system/mod.rs
@@ -1,5 +1,5 @@
 use crate::app_error::AppError;
-use crate::reddit_fetcher::reddit::ratelimit_headers::RatelimitHeaders;
+use crate::fetcher::reddit::ratelimit_headers::RatelimitHeaders;
 use crate::AppState;
 use axum::extract::State;
 use axum::Json;
