+++ backend/src/fetcher/model/mod.rs
