+++ backend/src/fetcher/model/posts.rs
@@ -1,9 +1,9 @@
 use crate::cast;
-use crate::reddit_fetcher::feed_request::{DataSource, FetcherFeedRequest};
-use crate::reddit_fetcher::fetcher_error::FetcherError;
-use crate::reddit_fetcher::model::reddit_data::RedditFeedData;
-use crate::reddit_fetcher::reddit::model::{RawContainer, RawPost};
-use crate::reddit_fetcher::reddit::request::SubredditPostsRequest;
+use crate::fetcher::feed_request::{DataSource, FetcherFeedRequest};
+use crate::fetcher::fetcher_error::FetcherError;
+use crate::fetcher::model::reddit_data::RedditFeedData;
+use crate::fetcher::reddit::model::{RawContainer, RawPost};
+use crate::fetcher::reddit::request::SubredditPostsRequest;
 use log_derive::logfn;
 use serde::Serialize;
