+++ backend/src/main.rs
@@ -1,7 +1,7 @@
+use crate::fetcher::fetcher::RMoodsFetcher;
+use crate::fetcher::reddit::connection::RedditConnection;
 use crate::nlp::nlp_client::NlpClient;
 use crate::open_api::ApiDoc;
-use crate::reddit_fetcher::fetcher::RMoodsFetcher;
-use crate::reddit_fetcher::reddit::connection::RedditConnection;
 use crate::startup::{shutdown_signal, verify_environment};
 use crate::websocket::SystemMessage;
 use axum::Router;
@@ -21,10 +21,10 @@ mod api;
 mod app_error;
 mod auth;
 mod env;
+mod fetcher;
 mod logging;
 mod nlp;
 mod open_api;
-mod reddit_fetcher;
 mod startup;
 mod websocket;
